refactors, renames and reorganize code from some files
