+++ aconcagua/bitboard.go
@@ -7,32 +7,6 @@ import (
 	"strings"
 )
 
-// files contains bitboards representing a file in chess board. Eg -> files[0] -> a File, files[1] -> b file, etc
-var files [8]Bitboard = [8]Bitboard{
-	0x0101010101010101,
-	0x0101010101010101 << 1,
-	0x0101010101010101 << 2,
-	0x0101010101010101 << 3,
-	0x0101010101010101 << 4,
-	0x0101010101010101 << 5,
-	0x0101010101010101 << 6,
-	0x0101010101010101 << 7,
-}
-
-// ranks contains bitboards representing a rank in chess board. Eg -> rank[0] -> rank 1, rank[1] -> rank 2, etc
-var ranks [8]Bitboard = [8]Bitboard{
-	0x00000000000000FF,
-	0x00000000000000FF << 8,
-	0x00000000000000FF << 16,
-	0x00000000000000FF << 24,
-	0x00000000000000FF << 32,
-	0x00000000000000FF << 40,
-	0x00000000000000FF << 48,
-	0x00000000000000FF << 56,
-}
-
-const AllSquares Bitboard = 0xFFFFFFFFFFFFFFFF
-
 // Bitboard represents a bitboard as a 64bit integer
 type Bitboard uint64
