+++ aconcagua/evaluation_test.go
@@ -5,8 +5,7 @@ import "testing"
 func TestEval(t *testing.T) {
 	pos := NewPosition()
 	pos.LoadFromFenString("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1")
-	pawnHashTable := NewPawnHashTable(1)
-	ev := pos.Evaluate(pawnHashTable)
+	ev := pos.Evaluate()
 
 	if ev != 0 {
 		t.Errorf("Expected: %v, got: %v", 0, ev)
