+++ aconcagua/move_generator.go
@@ -1,8 +1,209 @@
 package aconcagua
 
-// ------------------------------------------------------------------
-// PIECE MOVES GENERATION (BITBOARD)
-// ------------------------------------------------------------------
+const (
+	// Move Generation Stages flags
+	HashMoveStage = iota
+	GenerateCapturesStage
+	CapturesStage
+	FirstKillerStage
+	SecondKillerStage
+	// TODO: Counter move heruistic???
+	GenerateNonCapturesStage
+	NonCapturesStage
+	BadCapturesStage
+	EndStage
+)
+
+// MoveGenerator implements a staged move generator for a given position
+type MoveGenerator struct {
+	stage                              int
+	moveNumber                         int // the move count selected so far
+	pos                                *Position
+	pd                                 PositionData
+	hashMove                           *Move
+	killer1, killer2                   *Move
+	historyMoves                       *HistoryMovesTable
+	captures, nonCaptures, badCaptures MoveList
+}
+
+// NewMoveGenerator returns a new move generator
+func NewMoveGenerator(pos *Position, hashMove *Move, killer1 *Move, killer2 *Move, historyMoves *HistoryMovesTable) *MoveGenerator {
+	return &MoveGenerator{
+		stage:        HashMoveStage,
+		pos:          pos,
+		hashMove:     hashMove,
+		killer1:      killer1,
+		killer2:      killer2,
+		moveNumber:   -1, // NOTE: initialize with -1 to make the first move selected to have moveNumber = 0
+		captures:     NewMoveList(30),
+		badCaptures:  NewMoveList(30),
+		nonCaptures:  NewMoveList(100),
+		historyMoves: historyMoves,
+	}
+}
+
+// nextMove return the nextMove move of the position
+func (mg *MoveGenerator) nextMove() (move Move) {
+	mg.moveNumber++
+	switch mg.stage {
+	case HashMoveStage:
+		mg.stage = GenerateCapturesStage
+		if *mg.hashMove != NoMove {
+			return *mg.hashMove
+		}
+		fallthrough
+	case GenerateCapturesStage:
+		mg.pd = mg.pos.generatePositionData()
+		mg.stage = CapturesStage
+		mg.pos.generateCaptures(&mg.captures, &mg.pd)
+		scores := make([]int, len(mg.captures))
+		for i := range len(mg.captures) {
+			scores[i] = mg.pos.see(mg.captures[i].from(), mg.captures[i].to())
+		}
+		mg.captures.sort(scores)
+
+		// BadCaptures
+		for i := range scores {
+			if scores[i] < 0 {
+				mg.badCaptures = mg.captures[i:]
+				mg.captures = mg.captures[:i]
+				break
+			}
+		}
+
+		fallthrough
+	case CapturesStage:
+		move = *mg.captures.pickFirst()
+		if move != NoMove && move == *mg.hashMove {
+			move = *mg.captures.pickFirst()
+		}
+		if move != NoMove {
+			return move
+		}
+		mg.stage = FirstKillerStage
+		fallthrough
+	case FirstKillerStage:
+		mg.stage = SecondKillerStage
+		move = *mg.killer1
+		// NOTE: we need to validate legality of killers for this position, because the may be for the same ply, but of another branch of the tree!!
+		mg.pos.generateNonCaptures(&mg.nonCaptures, &mg.pd)
+		if move != NoMove && move != *mg.hashMove && isLegalKiller(move, &mg.nonCaptures) {
+			return move
+		}
+		fallthrough
+	case SecondKillerStage:
+		mg.stage = GenerateNonCapturesStage
+		move = *mg.killer2
+		if move != NoMove && move != *mg.hashMove && isLegalKiller(move, &mg.nonCaptures) {
+			return move
+		}
+		fallthrough
+	case GenerateNonCapturesStage:
+		mg.stage = NonCapturesStage
+		scores := make([]int, len(mg.nonCaptures))
+		for i := range len(mg.nonCaptures) {
+			scores[i] = mg.historyMoves[mg.pos.Turn][mg.nonCaptures[i].from()][mg.nonCaptures[i].to()]
+		}
+		mg.nonCaptures.sort(scores)
+		fallthrough
+	case NonCapturesStage:
+		move = *mg.nonCaptures.pickFirst()
+
+		if move != NoMove && move == *mg.hashMove {
+			move = *mg.nonCaptures.pickFirst()
+		}
+
+		if move != NoMove && move == *mg.killer1 {
+			move = *mg.nonCaptures.pickFirst()
+		}
+
+		if move != NoMove && move == *mg.killer2 {
+			move = *mg.nonCaptures.pickFirst()
+		}
+
+		if move != NoMove {
+			return move
+		}
+		mg.stage = BadCapturesStage
+		fallthrough
+	case BadCapturesStage:
+		move = *mg.badCaptures.pickFirst()
+		if move != NoMove && move == *mg.hashMove {
+			move = *mg.badCaptures.pickFirst()
+		}
+		if move != NoMove {
+			return move
+		}
+		mg.stage = EndStage
+	case EndStage:
+		return NoMove
+	}
+	return
+}
+
+// isLegalKiller returns if the move is legal in the current position
+func isLegalKiller(move Move, ml *MoveList) bool {
+	// Killer moves are always quiet moves, so we can just pass the non captures list to check if killer exits
+	for i := range len(*ml) {
+		if move == (*ml)[i] {
+			return true
+		}
+	}
+	return false
+}
+
+// generateCaptures generates all captures in the position and stores them in the move list
+func (pos *Position) generateCaptures(ml *MoveList, pd *PositionData) {
+	bitboards := pos.getBitboards(pos.Turn)
+
+	for piece, bb := range bitboards {
+		for bb > 0 {
+			pieceBB := bb.NextBit()
+			switch piece {
+			case King:
+				genMovesFromTargets(&pieceBB, kingMoves(&pieceBB, pos, pos.Turn)&pd.enemies, ml, pd)
+			case Queen:
+				genMovesFromTargets(&pieceBB, (rookMoves(&pieceBB, pd)|bishopMoves(&pieceBB, pd))&pd.enemies, ml, pd)
+			case Rook:
+				genMovesFromTargets(&pieceBB, rookMoves(&pieceBB, pd)&pd.enemies, ml, pd)
+			case Bishop:
+				genMovesFromTargets(&pieceBB, bishopMoves(&pieceBB, pd)&pd.enemies, ml, pd)
+			case Knight:
+				genMovesFromTargets(&pieceBB, knightMoves(&pieceBB, pd)&pd.enemies, ml, pd)
+			case Pawn:
+				genPawnCapturesMoves(&pieceBB, pos.Turn, ml, pd)
+			}
+		}
+	}
+	// TODO: add en passant captures here??. Need to fix see first(because 'to' square is an empty square)
+}
+
+// generateNonCaptures generates all non captures in the position and stores them in the move list
+func (pos *Position) generateNonCaptures(ml *MoveList, pd *PositionData) {
+	bitboards := pos.getBitboards(pos.Turn)
+
+	for piece, bb := range bitboards {
+		for bb > 0 {
+			pieceBB := bb.NextBit()
+			switch piece {
+			case King:
+				genMovesFromTargets(&pieceBB, kingMoves(&pieceBB, pos, pos.Turn)&^pd.enemies, ml, pd)
+				genCastleMoves(pos, ml)
+			case Queen:
+				genMovesFromTargets(&pieceBB, (rookMoves(&pieceBB, pd)|bishopMoves(&pieceBB, pd))&^pd.enemies, ml, pd)
+			case Rook:
+				genMovesFromTargets(&pieceBB, rookMoves(&pieceBB, pd)&^pd.enemies, ml, pd)
+			case Bishop:
+				genMovesFromTargets(&pieceBB, bishopMoves(&pieceBB, pd)&^pd.enemies, ml, pd)
+			case Knight:
+				genMovesFromTargets(&pieceBB, knightMoves(&pieceBB, pd)&^pd.enemies, ml, pd)
+			case Pawn:
+				genPawnMovesFromTarget(&pieceBB, pawnMoves(&pieceBB, pd, pos.Turn)&^pd.enemies, pos.Turn, ml, pd)
+			}
+		}
+	}
+	genEnPassantCaptures(pos, pos.Turn, ml)
+}
 
 // kingMoves returns a bitboard with the legal moves of the king from the bitboard passed
 func kingMoves(k *Bitboard, pos *Position, side Color) (moves Bitboard) {
@@ -82,10 +283,6 @@ func lastRank(side Color) (rank Bitboard) {
 	}
 	return
 }
-
-// ------------------------------------------------------------------
-// PIECE MOVES GENERATION (MOVE LIST)
-// ------------------------------------------------------------------
 
 // genMovesFromTargets generates the moves from the square passed to the targets passed in a MoveList
 func genMovesFromTargets(from *Bitboard, targets Bitboard, ml *MoveList, pd *PositionData) {
@@ -190,10 +387,6 @@ func genEnPassantCaptures(pos *Position, side Color, ml *MoveList) {
 	}
 }
 
-// ------------------------------------------------------------------
-// LEGAL MOVE VALIDATION FUNCTIONS
-// ------------------------------------------------------------------
-
 // checkRestrictedSquares returns a bitboard with the squares that are allowed to move when in check
 func checkRestrictedSquares(king Bitboard, checkingSliders Bitboard, checkingNonSliders Bitboard) (allowedSquares Bitboard) {
 	checkingPieces := checkingSliders | checkingNonSliders
@@ -219,57 +412,4 @@ func pinRestrictedSquares(piece Bitboard, king Bitboard, pinnedPieces Bitboard) 
 		return raysDirection(king, direction)
 	}
 	return AllSquares
-}
-
-// generateCaptures generates all captures in the position and stores them in the move list
-func (pos *Position) generateCaptures(ml *MoveList, pd *PositionData) {
-	bitboards := pos.getBitboards(pos.Turn)
-
-	for piece, bb := range bitboards {
-		for bb > 0 {
-			pieceBB := bb.NextBit()
-			switch piece {
-			case King:
-				genMovesFromTargets(&pieceBB, kingMoves(&pieceBB, pos, pos.Turn)&pd.enemies, ml, pd)
-			case Queen:
-				genMovesFromTargets(&pieceBB, (rookMoves(&pieceBB, pd)|bishopMoves(&pieceBB, pd))&pd.enemies, ml, pd)
-			case Rook:
-				genMovesFromTargets(&pieceBB, rookMoves(&pieceBB, pd)&pd.enemies, ml, pd)
-			case Bishop:
-				genMovesFromTargets(&pieceBB, bishopMoves(&pieceBB, pd)&pd.enemies, ml, pd)
-			case Knight:
-				genMovesFromTargets(&pieceBB, knightMoves(&pieceBB, pd)&pd.enemies, ml, pd)
-			case Pawn:
-				genPawnCapturesMoves(&pieceBB, pos.Turn, ml, pd)
-			}
-		}
-	}
-	// TODO: add en passant captures here??. Need to fix see first(because 'to' square is an empty square)
-}
-
-// generateNonCaptures generates all non captures in the position and stores them in the move list
-func (pos *Position) generateNonCaptures(ml *MoveList, pd *PositionData) {
-	bitboards := pos.getBitboards(pos.Turn)
-
-	for piece, bb := range bitboards {
-		for bb > 0 {
-			pieceBB := bb.NextBit()
-			switch piece {
-			case King:
-				genMovesFromTargets(&pieceBB, kingMoves(&pieceBB, pos, pos.Turn)&^pd.enemies, ml, pd)
-				genCastleMoves(pos, ml)
-			case Queen:
-				genMovesFromTargets(&pieceBB, (rookMoves(&pieceBB, pd)|bishopMoves(&pieceBB, pd))&^pd.enemies, ml, pd)
-			case Rook:
-				genMovesFromTargets(&pieceBB, rookMoves(&pieceBB, pd)&^pd.enemies, ml, pd)
-			case Bishop:
-				genMovesFromTargets(&pieceBB, bishopMoves(&pieceBB, pd)&^pd.enemies, ml, pd)
-			case Knight:
-				genMovesFromTargets(&pieceBB, knightMoves(&pieceBB, pd)&^pd.enemies, ml, pd)
-			case Pawn:
-				genPawnMovesFromTarget(&pieceBB, pawnMoves(&pieceBB, pd, pos.Turn)&^pd.enemies, pos.Turn, ml, pd)
-			}
-		}
-	}
-	genEnPassantCaptures(pos, pos.Turn, ml)
 }
