+++ aconcagua/move_generator_test.go
@@ -2,6 +2,91 @@ package aconcagua
 
 import "testing"
 
+func TestMoveGeneratorHasNextMove(t *testing.T) {
+	pos := NewPosition()
+	hashMove := encodeMove(0, 0, quiet)
+	killers := Killer{NoMove, NoMove}
+	mg := NewMoveGenerator(pos, hashMove, &killers[0], &killers[1], &HistoryMovesTable{})
+
+	got := mg.nextMove()
+
+	if got != NoMove {
+		t.Errorf("Expected: %v, got: %v", NoMove, got)
+	}
+}
+
+func TestMoveGeneratorNotHasNextMove(t *testing.T) {
+	pos := NewPosition()
+	hashMove := NoMove
+	killers := Killer{NoMove, NoMove}
+	mg := NewMoveGenerator(pos, &hashMove, &killers[0], &killers[1], &HistoryMovesTable{})
+	mg.stage = EndStage
+
+	expected := NoMove
+	got := mg.nextMove()
+
+	if got != expected {
+		t.Errorf("Expected: %v, got: %v", expected, got)
+	}
+}
+
+func TestMoveGeneratorCreatesCaptures(t *testing.T) {
+	pos := NewPosition()
+	pos.LoadFromFenString("1b4k1/5pp1/3r3p/4P3/5PN1/3RK3/8/8 w - - 0 1") // Only 3 captures
+	mg := NewMoveGenerator(pos, nil, nil, nil, nil)
+	move := NoMove
+	mg.hashMove = &move
+
+	expected := *encodeMove(36, 43, capture) // Best capture. Pawn takes rook
+	got := mg.nextMove()
+
+	if got != expected {
+		t.Errorf("Expected: %v, got: %v", expected, got)
+	}
+}
+
+func TestMoveGeneratorCreatesNonCaptures(t *testing.T) {
+	pos := NewPosition()
+	pos.LoadFromFenString("1b4k1/5pp1/3r3p/4P3/5PN1/3RK3/8/8 w - - 0 1") // Only 3 captures
+	noMove := NoMove
+	mg := NewMoveGenerator(pos, &noMove, &noMove, &noMove, &HistoryMovesTable{})
+	mg.stage = FirstKillerStage // NOTE: Non captures are generated in killers stage to validate legaliy of killers
+	move := NoMove
+	mg.hashMove = &move
+
+	mg.pd = mg.pos.generatePositionData()
+	mg.nextMove()
+
+	expected := 16 // 17 - 1 (the one picked)
+	got := len(mg.nonCaptures)
+
+	if got != expected {
+		t.Errorf("Expected: %v, got: %v", expected, got)
+	}
+}
+
+func TestMoveGeneratorGetsAllMoves(t *testing.T) {
+	pos := NewPosition()
+	pos.LoadFromFenString("1b4k1/5pp1/3r3p/4P3/5PN1/3RK3/8/8 w - - 0 1") // 3 captures + 17 non capt
+	hashMove := NoMove
+
+	ml := NewMoveList(20)
+	pd := pos.generatePositionData()
+	pos.generateNonCaptures(&ml, &pd)
+	killers := Killer{ml[0], ml[5]}
+
+	mg := NewMoveGenerator(pos, &hashMove, &killers[0], &killers[1], &HistoryMovesTable{})
+	for mg.nextMove() != NoMove {
+	}
+
+	expected := 20
+	got := mg.moveNumber
+
+	if got != expected {
+		t.Errorf("Expected: %v, got: %v", expected, got)
+	}
+}
+
 func TestStraightPinnedPieces(t *testing.T) {
 	pos := NewPosition()
 	pos.AddPiece(WhiteKing, "e1")
