+++ aconcagua/position.go
@@ -87,39 +87,7 @@ type Position struct {
 	halfmoveClock   int
 	fullmoveNumber  int
 	positionHistory PositionHistory
-	evaluation      Evaluation
-}
-
-// canNullMove returns if the current position allows a null move pruning
-func (pos *Position) canNullMove() bool {
-	if pos.material(pos.Turn) < EndgameMaterialThreshold {
-		return false
-	}
-
-	if pos.kingAndPawnsOnlyEndgame() {
-		return false
-	}
-
-	return true
-}
-
-// material returns the total material of the position for the side passed
-func (pos *Position) material(side Color) int {
-	pieceValue := [6]int{0, 900, 500, 300, 300, 100}
-	material := 0
-
-	for piece, bitboard := range pos.getBitboards(side) {
-		material += pieceValue[pieceRole(piece)] * bitboard.count()
-	}
-	return material
-}
-
-// kingAndPawnsOnlyEndgame returns if the position is a king and pawns only endgame
-func (pos *Position) kingAndPawnsOnlyEndgame() bool {
-	whiteKingAndPawns := pos.Bitboards[WhiteKing] | pos.Bitboards[WhitePawn]
-	blackKingAndPawns := pos.Bitboards[BlackKing] | pos.Bitboards[BlackPawn]
-
-	return pos.pieces[White] == whiteKingAndPawns && pos.pieces[Black] == blackKingAndPawns
+	eval            Evaluation
 }
 
 // PositionData contains relevant data for legal move validations of a position
@@ -145,7 +113,7 @@ func (pos *Position) generatePositionData() PositionData {
 	}
 }
 
-// PieceAt returns a Piece at the given square coordinate in the Position or error
+// PieceAt returns a Piece at the given square coordinate in the Position or NoPiece
 func (pos *Position) PieceAt(square string) (piece int) {
 	bitboardSquare := bitboardFromCoordinates(square)
 
@@ -239,11 +207,8 @@ func (pos *Position) Check(side Color) bool {
 		}
 	}
 
-	if pawnAttacks(&pos.Bitboards[pieceColor(Pawn, opponent)], opponent)&kingPos > 0 {
-		return true
-	}
-
-	return false
+	pawnCheck := pawnAttacks(&pos.Bitboards[pieceColor(Pawn, opponent)], opponent)&kingPos > 0
+	return pawnCheck
 }
 
 // pinnedPieces returns a bitboard with the pieces pinned in the position for the side passed
@@ -666,7 +631,7 @@ func toRuneArray(pos *Position) [64]rune {
 	return squares
 }
 
-// NewPositionFromFen creates a new Position struct from a fen string
+// LoadFromFenString creates a new Position struct from a fen string
 func (pos *Position) LoadFromFenString(fen string) {
 	pieceReference := map[string]int{
 		"k": BlackKing, "q": BlackQueen, "r": BlackRook, "b": BlackBishop, "n": BlackKnight, "p": BlackPawn,
@@ -713,7 +678,7 @@ func (pos *Position) LoadFromFenString(fen string) {
 	pos.Hash = zobristHashKeys.fullZobristHash(pos)
 	pos.PawnHash = zobristHashKeys.pawnHash(pos)
 
-	pos.evaluation.clear()
+	pos.eval.clear()
 }
 
 // NewPosition returns a new Position struct
@@ -721,6 +686,6 @@ func NewPosition() (pos *Position) {
 	return &Position{
 		positionHistory: *NewPositionHistory(),
 		castling:        *NewCastling(4, 7, 0),
-		evaluation:      *NewEvaluation(),
+		eval:            *NewEvaluation(DefaultPawnHashTableSizeInMb),
 	}
 }
