+++ main.go
@@ -10,11 +10,13 @@ func main() {
 	eng := aconcagua.NewEngine()
 	eng.StartUci()
 
+	// Tune Evaluation Params - Using zurich chess training set
 	// dataSet := tuner.LoadDataSet("./tuner/training-set/quiet-labeled.epd")
-	// params := tuner.GetEvaluationParams()
 
 	// k := tuner.FindOptimalScalingFactor(dataSet, tuner.GetEvaluationParams())
 	// fmt.Println("Best K", k)
 
-	// tuner.AdamTuner(params, dataSet, tuner.ScalingFactor, 2000)
+	// params := tuner.GetEvaluationParams()
+
+	// tuner.AdamTuner(params, dataSet, k, 2000)
 }
